+++ gamevault/UserControls/LibraryUserControl.xaml.cs
@@ -30,6 +30,30 @@ namespace gamevault.UserControls
         {
             InitializeComponent();
             ViewModel = new LibraryViewModel();
+            int sortByIndex = 0;
+            if (SettingsViewModel.Instance.RetainLibarySortByAndOrderBy)
+            {
+                try
+                {
+                    uiFilterOrderBy.IsChecked = Preferences.Get(AppConfigKey.LastLibraryOrderBy, AppFilePath.UserFile) == "desc" ? true : false;
+
+                    string lastSortBy = Preferences.Get(AppConfigKey.LastLibrarySortBy, AppFilePath.UserFile);
+
+                    for (int i = 0; i < ViewModel.GameFilterSortByValues.Count; i++)
+                    {
+                        if (ViewModel.GameFilterSortByValues.ElementAt(i).Value == lastSortBy)
+                        {
+                            sortByIndex = i;
+                            break;
+                        }
+                    }
+                }
+                catch { }
+            }
+            uiFilterSortBy.SelectedIndex = sortByIndex;
+            uiFilterOrderBy.Checked += OrderBy_Changed;
+            uiFilterOrderBy.Unchecked += OrderBy_Changed;
+
             this.DataContext = ViewModel;
             InitTimer();
         }
@@ -89,7 +113,7 @@ namespace gamevault.UserControls
             TaskQueue.Instance.ClearQueue();
 
             string gameSortByFilter = ViewModel.SelectedGameFilterSortBy.Value;
-            string gameOrderByFilter = ViewModel.OrderByValue;
+            string gameOrderByFilter = (bool)uiFilterOrderBy.IsChecked ? "DESC" : "ASC";
             ViewModel.GameCards.Clear();
             string filterUrl = @$"{SettingsViewModel.Instance.ServerUrl}/api/games?search={inputTimer.Data}&sortBy={gameSortByFilter}:{gameOrderByFilter}&limit=50";
             filterUrl = ApplyFilter(filterUrl);
@@ -252,7 +276,7 @@ namespace gamevault.UserControls
 
         private async void OrderBy_Changed(object sender, RoutedEventArgs e)
         {
-            ViewModel.OrderByValue = (ViewModel.OrderByValue == "ASC") ? "DESC" : "ASC";
+            Preferences.Set(AppConfigKey.LastLibraryOrderBy, (bool)uiFilterOrderBy.IsChecked ? "desc" : "asc", AppFilePath.UserFile);
             await Search();
         }
         private string ApplyFilter(string filter)
@@ -293,6 +317,10 @@ namespace gamevault.UserControls
         }
         private async void FilterUpdated(object sender, EventArgs e)
         {
+            if (sender == uiFilterSortBy)
+            {
+                Preferences.Set(AppConfigKey.LastLibrarySortBy, ViewModel.SelectedGameFilterSortBy.Value, AppFilePath.UserFile);
+            }
             OpenFilterIfClosed();
             RefreshFilterCounter();
             await Search();
@@ -330,7 +358,7 @@ namespace gamevault.UserControls
                 ContentControl parent = ((Grid)((FrameworkElement)sender).Parent).TemplatedParent as ContentControl;
                 if (parent.Tag == "busy")
                     return;
-                
+
                 parent.Tag = "busy";
                 try
                 {